tests/wpt/tests/workers/interfaces/WorkerGlobalScope/close/sending-messages.html

<!doctype html>
<title>close() and sending messages</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('sending-messages.js');
  var i = 0;
  worker.onmessage = this.step_func(function(e) {
    i++;
    assert_equals(e.data, i);
    if (i == 2) {
      this.done();
    }
  });
});
</script>
